Dont inline tRPC methods
