+++ server/routers/delete-card.ts
@@ -1,0 +1,31 @@
+import { z } from "zod";
+import { prismaClient } from "../prisma-client";
+import { procedure } from "../trpc";
+
+export const deleteCard = procedure
+  .input(
+    z.object({
+      id: z.optional(z.number()),
+    }),
+  )
+  .mutation(async ({ input, ctx }) => {
+    const userId = ctx.user?.id;
+    if (!userId) {
+      throw new Error("User not found");
+    }
+
+    const card = await prismaClient.card.findFirst({
+      where: {
+        id: input.id,
+        userId,
+      },
+    });
+
+    if (!card) {
+      throw new Error("Card not found");
+    }
+
+    await prismaClient.card.delete({
+      where: { id: card.id },
+    });
+  });
